first phase of fixed point arithmetic optimization

+++ fft.h
@@ -3,8 +3,8 @@
 
 typedef struct complex_t complex_t;
 struct complex_t{
-	float re;
-	float im;
+	unsigned int re;
+	unsigned int im;
 };
 
 
@@ -17,8 +17,8 @@ void swap_complex_array(complex_t* c, int n);
 //complex_t twiddle(int N,int k);
 void complex_mult(complex_t* left, complex_t* right, complex_t* rs);
 void complex_add(complex_t* left, complex_t* right, complex_t* rs);
-int forward_fft(complex_t* input, complex_t* output, int n);
-int inverse_fft(complex_t* input, complex_t* output, int n);
+int forward_fft(complex_t* input, complex_t* output, unsigned n);
+int inverse_fft(complex_t* input, complex_t* output, unsigned n);
 
 // input: pointer to an array of complex input samples
 // ouptut: pointer to an array where the result can be placed.
@@ -28,6 +28,6 @@ int inverse_fft(complex_t* input, complex_t* output, int n);
 //		input and output are of size n
 // 	n must be a power of 2.
 // DFT is defined as X(n) = sum( x[n] * e^(-2jnk*pi/N ) for n = 0 to N-1
-int _fft(complex_t* input, complex_t* output, int n);
+int _fft(complex_t* input, complex_t* output, unsigned n);
 
 #endif
